XW-1774 | rename prepare-media-wiki-data.js file to match the function name
